feature/ELK-610-database-models-revision-1
